requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
MouseInfo==0.1.2
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyAutoGUI==0.9.48
pycparser==2.19
PyGetWindow==0.0.7
pylint==2.3.1
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.24
python-dotenv==0.10.3
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.1
qiniu==7.2.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
whitenoise==4.1.3
wrapt==1.11.2
